Stage configuration moved to Screen.configure(), which also now saves the actual screen width and height presuming it's called before any other operations are conducted.
Mouse was reworked a bit to work around hiding bug in Chromium.

GameLoop was cleaned up a bit.  It and World were changed to make more
clear that they hold world dimensions and display dimensions, the
latter of which is not necessarily the same as screen dimensions.

Both samples were updated to reflect these changes.

The new tests/00-ScreenDimension is just for playing around with how
changes in scale mode alter reported stage dimensions.